Tối 18/9

+++ app/src/main/java/com/bkav/mymusic/Song.java
@@ -13,7 +13,7 @@ public class Song {
     private int duration;
 
 
-    public long getId() {
+    public int getId() {
         return id;
     }
